refactor: create files for item components and import in home page
